+++ PythonAPI/agents/navigation/basic_agent.py
@@ -10,18 +10,12 @@
 waypoints and avoiding other vehicles.
 The agent also responds to traffic lights. """
 
-import math
-
-import numpy as np
 
 import carla
 from agents.navigation.agent import Agent, AgentState
 from agents.navigation.local_planner import LocalPlanner
-from agents.navigation.local_planner import RoadOption
 from agents.navigation.global_route_planner import GlobalRoutePlanner
 from agents.navigation.global_route_planner_dao import GlobalRoutePlannerDAO
-from agents.tools.misc import vector
-
 
 class BasicAgent(Agent):
     """
@@ -38,93 +32,54 @@ class BasicAgent(Agent):
 
         self._proximity_threshold = 10.0  # meters
         self._state = AgentState.NAVIGATING
-        self._local_planner = LocalPlanner(self._vehicle, opt_dict={'target_speed': target_speed})
+        args_lateral_dict = {
+            'K_P': 1,
+            'K_D': 0.02,
+            'K_I': 0,
+            'dt': 1.0/20.0}
+        self._local_planner = LocalPlanner(
+            self._vehicle, opt_dict={'target_speed' : target_speed,
+            'lateral_control_dict':args_lateral_dict})
         self._hop_resolution = 2.0
-
-        # setting up global router
-        self._current_plan = None
+        self._path_seperation_hop = 2
+        self._path_seperation_threshold = 0.5
+        self._target_speed = target_speed
+        self._grp = None
 
     def set_destination(self, location):
+        """
+        This method creates a list of waypoints from agent's position to destination location
+        based on the route returned by the global router
+        """
+
         start_waypoint = self._map.get_waypoint(self._vehicle.get_location())
         end_waypoint = self._map.get_waypoint(
             carla.Location(location[0], location[1], location[2]))
-        solution = []
+
+        route_trace = self._trace_route(start_waypoint, end_waypoint)
+        assert route_trace
+
+        self._local_planner.set_global_plan(route_trace)
+
+    def _trace_route(self, start_waypoint, end_waypoint):
+        """
+        This method sets up a global router and returns the optimal route
+        from start_waypoint to end_waypoint
+        """
 
         # Setting up global router
-        dao = GlobalRoutePlannerDAO(self._vehicle.get_world().get_map())
-        grp = GlobalRoutePlanner(dao)
-        grp.setup()
+        if self._grp is None:
+            dao = GlobalRoutePlannerDAO(self._vehicle.get_world().get_map(), self._hop_resolution)
+            grp = GlobalRoutePlanner(dao)
+            grp.setup()
+            self._grp = grp
 
         # Obtain route plan
-        x1 = start_waypoint.transform.location.x
-        y1 = start_waypoint.transform.location.y
-        x2 = end_waypoint.transform.location.x
-        y2 = end_waypoint.transform.location.y
-        route = grp.plan_route((x1, y1), (x2, y2))
+        route = self._grp.trace_route(
+            start_waypoint.transform.location,
+            end_waypoint.transform.location)
 
-        current_waypoint = start_waypoint
-        route.append(RoadOption.VOID)
-        for action in route:
-
-            #   Generate waypoints to next junction
-            wp_choice = current_waypoint.next(self._hop_resolution)
-            while len(wp_choice) == 1:
-                current_waypoint = wp_choice[0]
-                solution.append((current_waypoint, RoadOption.LANEFOLLOW))
-                wp_choice = current_waypoint.next(self._hop_resolution)
-                #   Stop at destination
-                if current_waypoint.transform.location.distance(
-                        end_waypoint.transform.location) < self._hop_resolution:
-                    break
-            if action == RoadOption.VOID:
-                break
-
-            #   Select appropriate path at the junction
-            if len(wp_choice) > 1:
-
-                # Current heading vector
-                current_transform = current_waypoint.transform
-                current_location = current_transform.location
-                projected_location = current_location + \
-                    carla.Location(
-                        x=math.cos(math.radians(current_transform.rotation.yaw)),
-                        y=math.sin(math.radians(current_transform.rotation.yaw)))
-                v_current = vector(current_location, projected_location)
-
-                direction = 0
-                if action == RoadOption.LEFT:
-                    direction = 1
-                elif action == RoadOption.RIGHT:
-                    direction = -1
-                elif action == RoadOption.STRAIGHT:
-                    direction = 0
-                select_criteria = float('inf')
-
-                #   Choose correct path
-                for wp_select in wp_choice:
-                    v_select = vector(
-                        current_location, wp_select.transform.location)
-                    cross = float('inf')
-                    if direction == 0:
-                        cross = abs(np.cross(v_current, v_select)[-1])
-                    else:
-                        cross = direction * np.cross(v_current, v_select)[-1]
-                    if cross < select_criteria:
-                        select_criteria = cross
-                        current_waypoint = wp_select
-
-                #   Generate all waypoints within the junction
-                #   along selected path
-                solution.append((current_waypoint, action))
-                current_waypoint = current_waypoint.next(self._hop_resolution)[0]
-                while current_waypoint.is_intersection:
-                    solution.append((current_waypoint, action))
-                    current_waypoint = current_waypoint.next(self._hop_resolution)[0]
-
-        assert solution
-
-        self._current_plan = solution
-        self._local_planner.set_global_plan(self._current_plan)
+        return route
 
     def run_step(self, debug=False):
         """