+++ PythonAPI/agents/navigation/global_route_planner_dao.py
@@ -5,6 +5,8 @@
 This module provides implementation for GlobalRoutePlannerDAO
 """
 
+import numpy as np
+
 
 class GlobalRoutePlannerDAO(object):
     """
@@ -12,12 +14,13 @@ class GlobalRoutePlannerDAO(object):
     from the carla server instance for GlobalRoutePlanner
     """
 
-    def __init__(self, wmap):
-        """
+    def __init__(self, wmap, sampling_resolution=1):
+        """get_topology
         Constructor
 
         wmap    :   carl world map object
         """
+        self._sampling_resolution = sampling_resolution
         self._wmap = wmap
 
     def get_topology(self):
@@ -28,34 +31,43 @@ class GlobalRoutePlannerDAO(object):
         topology into a list of dictionary objects.
 
         return: list of dictionary objects with the following attributes
-                entry   -   (x,y) of entry point of road segment
-                exit    -   (x,y) of exit point of road segment
+                entry   -   waypoint of entry point of road segment
+                entryxyz-   (x,y,z) of entry point of road segment
+                exit    -   waypoint of exit point of road segment
+                exitxyz -   (x,y,z) of exit point of road segment
                 path    -   list of waypoints separated by 1m from entry
                             to exit
-                intersection    -   Boolean indicating if the road segment
-                                    is an intersection
         """
         topology = []
         # Retrieving waypoints to construct a detailed topology
         for segment in self._wmap.get_topology():
-            x1 = segment[0].transform.location.x
-            y1 = segment[0].transform.location.y
-            x2 = segment[1].transform.location.x
-            y2 = segment[1].transform.location.y
+            wp1, wp2 = segment[0], segment[1]
+            l1, l2 = wp1.transform.location, wp2.transform.location
+            # Rounding off to avoid floating point imprecision
+            x1, y1, z1, x2, y2, z2 = np.round([l1.x, l1.y, l1.z, l2.x, l2.y, l2.z], 0)
+            wp1.transform.location, wp2.transform.location = l1, l2
             seg_dict = dict()
-            seg_dict['entry'] = (x1, y1)
-            seg_dict['exit'] = (x2, y2)
+            seg_dict['entry'], seg_dict['exit'] = wp1, wp2
+            seg_dict['entryxyz'], seg_dict['exitxyz'] = (x1, y1, z1), (x2, y2, z2)
             seg_dict['path'] = []
-            wp1 = segment[0]
-            wp2 = segment[1]
-            seg_dict['intersection'] = True if wp1.is_intersection else False
             endloc = wp2.transform.location
-            w = wp1.next(1)[0]
-            while w.transform.location.distance(endloc) > 1:
-                x = w.transform.location.x
-                y = w.transform.location.y
-                seg_dict['path'].append((x, y))
-                w = w.next(1)[0]
-
+            if wp1.transform.location.distance(endloc) > self._sampling_resolution:
+                w = wp1.next(self._sampling_resolution)[0]
+                while w.transform.location.distance(endloc) > self._sampling_resolution:
+                    seg_dict['path'].append(w)
+                    w = w.next(self._sampling_resolution)[0]
+            else:
+                seg_dict['path'].append(wp1.next(self._sampling_resolution/2.0)[0])
             topology.append(seg_dict)
         return topology
+
+    def get_waypoint(self, location):
+        """
+        The method returns waypoint at given location
+        """
+        waypoint = self._wmap.get_waypoint(location)
+        return waypoint
+
+    def get_resolution(self):
+        """ Accessor for self._sampling_resolution """
+        return self._sampling_resolution