+++ PythonAPI/agents/navigation/local_planner.py
@@ -26,6 +26,8 @@ class RoadOption(Enum):
     RIGHT = 2
     STRAIGHT = 3
     LANEFOLLOW = 4
+    CHANGELANELEFT = 5
+    CHANGELANERIGHT = 6
 
 
 class LocalPlanner(object):
@@ -61,23 +63,23 @@ class LocalPlanner(object):
         self._vehicle = vehicle
         self._map = self._vehicle.get_world().get_map()
 
-        self._buffer_size = 5
-        self.dt = None
-        self.target_speed = None
-        self.sampling_radius = None
-        self.min_distance = None
-        self.current_waypoint = None
-        self.target_road_option = None
-        self.next_waypoints = None
+        self._dt = None
+        self._target_speed = None
+        self._sampling_radius = None
+        self._min_distance = None
+        self._current_waypoint = None
+        self._target_road_option = None
+        self._next_waypoints = None
         self.target_waypoint = None
-        self.vehicle_controller = None
-        self.global_plan = None
+        self._vehicle_controller = None
+        self._global_plan = None
         # queue with tuples of (waypoint, RoadOption)
-        self.waypoints_queue = deque(maxlen=600)
-        self.waypoint_buffer = deque(maxlen=self._buffer_size)
+        self._waypoints_queue = deque(maxlen=20000)
+        self._buffer_size = 5
+        self._waypoint_buffer = deque(maxlen=self._buffer_size)
 
         # initializing controller
-        self.init_controller(opt_dict)
+        self._init_controller(opt_dict)
 
     def __del__(self):
         if self._vehicle:
@@ -88,7 +90,7 @@ class LocalPlanner(object):
         self._vehicle = None
         print("Resetting ego-vehicle!")
 
-    def init_controller(self, opt_dict):
+    def _init_controller(self, opt_dict):
         """
         Controller initialization.
 
@@ -96,46 +98,46 @@ class LocalPlanner(object):
         :return:
         """
         # default params
-        self.dt = 1.0 / 20.0
-        self.target_speed = 20.0  # Km/h
-        self.sampling_radius = self.target_speed * 0.5 / 3.6  # 0.5 seconds horizon
-        self.min_distance = self.sampling_radius * self.MIN_DISTANCE_PERCENTAGE
+        self._dt = 1.0 / 20.0
+        self._target_speed = 20.0  # Km/h
+        self._sampling_radius = self._target_speed * 1 / 3.6  # 1 seconds horizon
+        self._min_distance = self._sampling_radius * self.MIN_DISTANCE_PERCENTAGE
         args_lateral_dict = {
             'K_P': 1.95,
             'K_D': 0.01,
             'K_I': 1.4,
-            'dt': self.dt}
+            'dt': self._dt}
         args_longitudinal_dict = {
             'K_P': 1.0,
             'K_D': 0,
             'K_I': 1,
-            'dt': self.dt}
+            'dt': self._dt}
 
         # parameters overload
         if opt_dict:
             if 'dt' in opt_dict:
-                self.dt = opt_dict['dt']
+                self._dt = opt_dict['dt']
             if 'target_speed' in opt_dict:
-                self.target_speed = opt_dict['target_speed']
+                self._target_speed = opt_dict['target_speed']
             if 'sampling_radius' in opt_dict:
-                self.sampling_radius = self.target_speed * \
+                self._sampling_radius = self._target_speed * \
                     opt_dict['sampling_radius'] / 3.6
             if 'lateral_control_dict' in opt_dict:
                 args_lateral_dict = opt_dict['lateral_control_dict']
             if 'longitudinal_control_dict' in opt_dict:
                 args_longitudinal_dict = opt_dict['longitudinal_control_dict']
 
-        self.current_waypoint = self._map.get_waypoint(self._vehicle.get_location())
-        self.vehicle_controller = VehiclePIDController(self._vehicle,
+        self._current_waypoint = self._map.get_waypoint(self._vehicle.get_location())
+        self._vehicle_controller = VehiclePIDController(self._vehicle,
                                                        args_lateral=args_lateral_dict,
                                                        args_longitudinal=args_longitudinal_dict)
 
-        self.global_plan = False
+        self._global_plan = False
 
         # compute initial waypoints
-        self.waypoints_queue.append((self.current_waypoint.next(self.sampling_radius)[0], RoadOption.LANEFOLLOW))
+        self._waypoints_queue.append((self._current_waypoint.next(self._sampling_radius)[0], RoadOption.LANEFOLLOW))
 
-        self.target_road_option = RoadOption.LANEFOLLOW
+        self._target_road_option = RoadOption.LANEFOLLOW
         # fill waypoint trajectory queue
         self._compute_next_waypoints(k=200)
 
@@ -146,7 +148,7 @@ class LocalPlanner(object):
         :param speed: new target speed in Km/h
         :return:
         """
-        self.target_speed = speed
+        self._target_speed = speed
 
     def _compute_next_waypoints(self, k=1):
         """
@@ -156,12 +158,12 @@ class LocalPlanner(object):
         :return:
         """
         # check we do not overflow the queue
-        available_entries = self.waypoints_queue.maxlen - len(self.waypoints_queue)
+        available_entries = self._waypoints_queue.maxlen - len(self._waypoints_queue)
         k = min(available_entries, k)
 
         for _ in range(k):
-            last_waypoint = self.waypoints_queue[-1][0]
-            next_waypoints = list(last_waypoint.next(self.sampling_radius))
+            last_waypoint = self._waypoints_queue[-1][0]
+            next_waypoints = list(last_waypoint.next(self._sampling_radius))
 
             if len(next_waypoints) == 1:
                 # only one option available ==> lanefollowing
@@ -169,20 +171,20 @@ class LocalPlanner(object):
                 road_option = RoadOption.LANEFOLLOW
             else:
                 # random choice between the possible options
-                road_options_list = retrieve_options(
+                road_options_list = _retrieve_options(
                     next_waypoints, last_waypoint)
                 road_option = random.choice(road_options_list)
                 next_waypoint = next_waypoints[road_options_list.index(
                     road_option)]
 
-            self.waypoints_queue.append((next_waypoint, road_option))
+            self._waypoints_queue.append((next_waypoint, road_option))
 
     def set_global_plan(self, current_plan):
-        self.waypoints_queue.clear()
+        self._waypoints_queue.clear()
         for elem in current_plan:
-            self.waypoints_queue.append(elem)
-        self.target_road_option = RoadOption.LANEFOLLOW
-        self.global_plan = True
+            self._waypoints_queue.append(elem)
+        self._target_road_option = RoadOption.LANEFOLLOW
+        self._global_plan = True
 
     def run_step(self, debug=True):
         """
@@ -194,47 +196,46 @@ class LocalPlanner(object):
         """
 
         # not enough waypoints in the horizon? => add more!
-        if len(self.waypoints_queue) < int(self.waypoints_queue.maxlen * 0.5):
-            if not self.global_plan:
-                self._compute_next_waypoints(k=100)
+        if not self._global_plan and len(self._waypoints_queue) < int(self._waypoints_queue.maxlen * 0.5):
+            self._compute_next_waypoints(k=100)
 
-        if len(self.waypoints_queue) == 0:
+        if len(self._waypoints_queue) == 0:
             control = carla.VehicleControl()
             control.steer = 0.0
             control.throttle = 0.0
-            control.brake = 0.0
+            control.brake = 1.0
             control.hand_brake = False
             control.manual_gear_shift = False
 
             return control
 
         #   Buffering the waypoints
-        if not self.waypoint_buffer:
+        if not self._waypoint_buffer:
             for i in range(self._buffer_size):
-                if self.waypoints_queue:
-                    self.waypoint_buffer.append(
-                        self.waypoints_queue.popleft())
+                if self._waypoints_queue:
+                    self._waypoint_buffer.append(
+                        self._waypoints_queue.popleft())
                 else:
                     break
 
         # current vehicle waypoint
-        self.current_waypoint = self._map.get_waypoint(self._vehicle.get_location())
+        self._current_waypoint = self._map.get_waypoint(self._vehicle.get_location())
         # target waypoint
-        self.target_waypoint, self.target_road_option = self.waypoint_buffer[0]
+        self.target_waypoint, self._target_road_option = self._waypoint_buffer[0]
         # move using PID controllers
-        control = self.vehicle_controller.run_step(self.target_speed, self.target_waypoint)
+        control = self._vehicle_controller.run_step(self._target_speed, self.target_waypoint)
 
         # purge the queue of obsolete waypoints
         vehicle_transform = self._vehicle.get_transform()
         max_index = -1
 
-        for i, (waypoint, _) in enumerate(self.waypoint_buffer):
+        for i, (waypoint, _) in enumerate(self._waypoint_buffer):
             if distance_vehicle(
-                    waypoint, vehicle_transform) < self.min_distance:
+                    waypoint, vehicle_transform) < self._min_distance:
                 max_index = i
         if max_index >= 0:
             for i in range(max_index + 1):
-                self.waypoint_buffer.popleft()
+                self._waypoint_buffer.popleft()
 
         if debug:
             draw_waypoints(self._vehicle.get_world(), [self.target_waypoint], self._vehicle.get_location().z + 1.0)
@@ -242,7 +243,7 @@ class LocalPlanner(object):
         return control
 
 
-def retrieve_options(list_waypoints, current_waypoint):
+def _retrieve_options(list_waypoints, current_waypoint):
     """
     Compute the type of connection between the current active waypoint and the multiple waypoints present in
     list_waypoints. The result is encoded as a list of RoadOption enums.
@@ -258,13 +259,13 @@ def retrieve_options(list_waypoints, current_waypoint):
         # the beggining of an intersection, therefore the
         # variation in angle is small
         next_next_waypoint = next_waypoint.next(3.0)[0]
-        link = compute_connection(current_waypoint, next_next_waypoint)
+        link = _compute_connection(current_waypoint, next_next_waypoint)
         options.append(link)
 
     return options
 
 
-def compute_connection(current_waypoint, next_waypoint):
+def _compute_connection(current_waypoint, next_waypoint):
     """
     Compute the type of topological connection between an active waypoint (current_waypoint) and a target waypoint
     (next_waypoint).